gradientMagnitudes is now of type uchar in EDL.

+++ algorithms/include/edl/edl.h
@@ -26,11 +26,11 @@ private:
 
     void routeAnchors(double angleTolerance, cv::InputArray magnitudes, cv::InputArray angles, std::vector<cv::Point2i*>& anchorPoints, std::vector<Line*>& result);
 
-    void walkFromAnchor(cv::Point& anchorPoint, double angleTolerance, cv::Mat_<double>& gradientMagnitudes, cv::Mat_<double>& gradientAngles, cv::Mat_<bool> &edgels, std::vector<std::vector<cv::Point*>*>& lineSegments);
+    void walkFromAnchor(cv::Point& anchorPoint, double angleTolerance, cv::Mat_<uchar>& gradientMagnitudes, cv::Mat_<double>& gradientAngles, cv::Mat_<bool> &edgels, std::vector<std::vector<cv::Point*>*>& lineSegments);
 
     bool getDirection(cv::Point& point, cv::Mat& gradientAngles);
 
-    cv::Point* findNextPoint(cv::Point* currentPoint, int mainDirection, int subDirection, cv::Mat_<double>& gradientMagnitudes);
+    cv::Point* findNextPoint(cv::Point* currentPoint, int mainDirection, int subDirection, cv::Mat_<uchar>& gradientMagnitudes);
 
     bool isAligned(double compare, double angle, double tolerance);
 };
